separate https_client into own files
