+++ aclk/aclk_common.h
@@ -3,8 +3,6 @@
 
 #include "libnetdata/libnetdata.h"
 
-#include <libwebsockets.h>
-
 typedef enum aclk_proxy_type {
     PROXY_TYPE_UNKNOWN = 0,
     PROXY_TYPE_SOCKS5,
@@ -21,7 +19,5 @@ ACLK_PROXY_TYPE aclk_verify_proxy(const char *string);
 const char *aclk_lws_wss_get_proxy_setting(ACLK_PROXY_TYPE *type);
 void safe_log_proxy_censor(char *proxy);
 int aclk_decode_base_url(char *url, char **aclk_hostname, char **aclk_port);
-
-int aclk_wss_set_socks(struct lws_vhost *vhost, const char *socks);
 
 #endif //ACLK_COMMON_H
