+++ aclk/aclk_lws_wss_client.h
@@ -1,8 +1,11 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+
 #ifndef ACLK_LWS_WSS_CLIENT_H
 #define ACLK_LWS_WSS_CLIENT_H
 
+#include <libwebsockets.h>
+
 #include "libnetdata/libnetdata.h"
-#include "aclk_common.h"
 
 // This is as define because ideally the ACLK at high level
 // can do mosqitto writes and reads only from one thread
@@ -77,6 +80,8 @@ void aclk_lws_connection_established();
 void aclk_lws_connection_data_received();
 void aclk_lws_connection_closed();
 void lws_wss_check_queues(size_t *write_len, size_t *write_len_bytes, size_t *read_len);
+
+int aclk_wss_set_socks(struct lws_vhost *vhost, const char *socks);
 
 #define FRAGMENT_SIZE 4096
 #endif
